refactor: rename connector package to plugin
