+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/GraviteeManagementV2Application.java
@@ -24,8 +24,8 @@ import io.gravitee.rest.api.management.v2.rest.resource.api.ApiMembersResource;
 import io.gravitee.rest.api.management.v2.rest.resource.api.ApiPlansResource;
 import io.gravitee.rest.api.management.v2.rest.resource.api.ApiResource;
 import io.gravitee.rest.api.management.v2.rest.resource.api.ApisResource;
-import io.gravitee.rest.api.management.v2.rest.resource.connector.EndpointsResource;
-import io.gravitee.rest.api.management.v2.rest.resource.connector.EntrypointsResource;
+import io.gravitee.rest.api.management.v2.rest.resource.plugin.EndpointsResource;
+import io.gravitee.rest.api.management.v2.rest.resource.plugin.EntrypointsResource;
 import io.gravitee.rest.api.management.v2.rest.resource.group.GroupResource;
 import io.gravitee.rest.api.management.v2.rest.resource.installation.EnvironmentsResource;
 import io.gravitee.rest.api.management.v2.rest.resource.installation.GraviteeLicenseResource;
