+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/io/gravitee/rest/api/management/v2/rest/resource/plugin/EntrypointsResourceTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.management.v2.rest.resource.connector;
+package io.gravitee.rest.api.management.v2.rest.resource.plugin;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.reset;
